added isEnabled column to User table
