+++ src/main/java/com/davistiba/wedemyserver/config/MainUserDetails.java
@@ -1,0 +1,70 @@
+package com.davistiba.wedemyserver.config;
+
+import com.davistiba.wedemyserver.models.User;
+import com.fasterxml.jackson.annotation.JsonIgnore;
+import lombok.Getter;
+import org.springframework.security.core.GrantedAuthority;
+import org.springframework.security.core.authority.SimpleGrantedAuthority;
+import org.springframework.security.core.userdetails.UserDetails;
+
+import java.io.Serial;
+import java.io.Serializable;
+import java.util.Collection;
+import java.util.Collections;
+
+@Getter
+public class MainUserDetails implements UserDetails, Serializable {
+
+    @Serial
+    private static final long serialVersionUID = 3733936374876008250L;
+
+    private final User user;
+
+    public MainUserDetails(User user) {
+        this.user = user;
+    }
+
+    @Override
+    @JsonIgnore
+    public Collection<? extends GrantedAuthority> getAuthorities() {
+        SimpleGrantedAuthority authority = new SimpleGrantedAuthority(user.getUserRole());
+        return Collections.singletonList(authority);
+    }
+
+    @Override
+    @JsonIgnore
+    public String getPassword() {
+        return user.getPassword();
+    }
+
+    @Override
+    @JsonIgnore
+    public String getUsername() {
+        return user.getEmail();
+    }
+
+    @Override
+    @JsonIgnore
+    public boolean isAccountNonExpired() {
+        return true;
+    }
+
+    @Override
+    @JsonIgnore
+    public boolean isAccountNonLocked() {
+        return true;
+    }
+
+    @Override
+    @JsonIgnore
+    public boolean isCredentialsNonExpired() {
+        return true;
+    }
+
+    @Override
+    @JsonIgnore
+    public boolean isEnabled() {
+        return user.getEnabled();
+    }
+
+}
